handle inbound-outbound route matching for sockets
