+++ main/src/net/http.rs
@@ -1,18 +1,14 @@
 use super::*;
 
-use crate::websocket::{is_websocket_req, spawn_proxy};
-
 use address_status::{
     get_address_status_list, post_address_status, AddressStatus, AddressStatusOperations,
 };
 use ctx::{AppConfig, ProxyRoute};
 use db::*;
 use hyper::header::HeaderName;
-use hyper::server::conn::AddrStream;
-use hyper::service::{make_service_fn, service_fn};
 use hyper::{
     header::{self, HeaderValue},
-    Body, HeaderMap, Method, Request, Response, Server, StatusCode,
+    Body, HeaderMap, Method, Request, Response, StatusCode,
 };
 use hyper_tls::HttpsConnector;
 use jwt::{get_cached_token_or_generate_one, JwtClaims};
@@ -21,27 +17,7 @@ use rate_limiter::RateLimitOperations;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use sign::SignedMessage;
-use std::net::{IpAddr, SocketAddr};
-use std::str::FromStr;
-
-macro_rules! http_log_format {
-  ($ip: expr, $address: expr, $path: expr, $format: expr, $($args: tt)+) => {format!(concat!("[Ip: {} | Address: {} | Path: {}] ", $format), $ip, $address, $path, $($args)+)};
-  ($ip: expr, $address: expr, $path: expr, $format: expr) => {format!(concat!("[Ip: {} | Pubkey: {} | Address: {}] ", $format), $ip, $address, $path)}
-}
-
-impl AppConfig {
-    pub(crate) fn get_proxy_route_by_inbound(&self, inbound: String) -> Option<&ProxyRoute> {
-        let route_index = self.proxy_routes.iter().position(|r| {
-            r.inbound_route == inbound || r.inbound_route.to_owned() + "/" == inbound
-        });
-
-        if let Some(index) = route_index {
-            return Some(&self.proxy_routes[index]);
-        }
-
-        None
-    }
-}
+use std::net::SocketAddr;
 
 async fn get_healthcheck() -> GenericResult<Response<Body>> {
     let json = json!({
@@ -119,7 +95,7 @@ async fn proxy(
     if !proxy_route.allowed_methods.contains(&payload.method) {
         log::warn!(
             "{}",
-            http_log_format!(
+            log_format!(
                 remote_addr.ip(),
                 payload.signed_message.address,
                 req.uri(),
@@ -137,7 +113,7 @@ async fn proxy(
     {
         log::error!(
             "{}",
-            http_log_format!(
+            log_format!(
                 remote_addr.ip(),
                 payload.signed_message.address,
                 req.uri(),
@@ -153,7 +129,7 @@ async fn proxy(
         Err(_) => {
             log::error!(
                 "{}",
-                http_log_format!(
+                log_format!(
                     remote_addr.ip(),
                     payload.signed_message.address,
                     original_req_uri,
@@ -195,7 +171,7 @@ async fn proxy(
         Err(_) => {
             log::warn!(
                 "{}",
-                http_log_format!(
+                log_format!(
                     remote_addr.ip(),
                     payload.signed_message.address,
                     original_req_uri,
@@ -210,50 +186,17 @@ async fn proxy(
     Ok(res)
 }
 
-#[allow(dead_code)]
-fn get_real_address(req: &Request<Body>, remote_addr: &SocketAddr) -> GenericResult<SocketAddr> {
-    if let Some(ip) = req.headers().get("x-forwarded-for") {
-        let addr = IpAddr::from_str(ip.to_str()?)?;
-
-        return Ok(SocketAddr::new(addr, remote_addr.port()));
-    }
-
-    Ok(*remote_addr)
-}
-
-async fn router(
+pub(crate) async fn http_handler(
     cfg: &AppConfig,
     req: Request<Body>,
     remote_addr: SocketAddr,
 ) -> GenericResult<Response<Body>> {
-    let remote_addr = match get_real_address(&req, &remote_addr) {
-        Ok(t) => t,
-        _ => {
-            log::error!(
-                "{}",
-                http_log_format!(
-                    remote_addr.ip(),
-                    String::from("-"),
-                    req.uri(),
-                    "Reading real remote address failed, returning 500."
-                )
-            );
-            return response_by_status(StatusCode::INTERNAL_SERVER_ERROR);
-        }
-    };
-
-    if is_websocket_req(&req) {
-        // TODO
-        // Spawn it
-        return spawn_proxy(req).await;
-    }
-
     let req_uri = req.uri().clone();
 
     if !remote_addr.ip().is_global() {
         log::info!(
             "{}",
-            http_log_format!(
+            log_format!(
                 remote_addr.ip(),
                 String::from("-"),
                 req.uri(),
@@ -278,7 +221,7 @@ async fn router(
         Err(_) => {
             log::warn!(
                 "{}",
-                http_log_format!(
+                log_format!(
                     remote_addr.ip(),
                     String::from("-"),
                     req_uri,
@@ -291,7 +234,7 @@ async fn router(
 
     log::info!(
         "{}",
-        http_log_format!(
+        log_format!(
             remote_addr.ip(),
             payload.signed_message.address,
             req_uri,
@@ -304,7 +247,7 @@ async fn router(
         Err(_) => {
             log::error!(
                 "{}",
-                http_log_format!(
+                log_format!(
                     remote_addr.ip(),
                     payload.signed_message.address,
                     req_uri,
@@ -320,7 +263,7 @@ async fn router(
         None => {
             log::warn!(
                 "{}",
-                http_log_format!(
+                log_format!(
                     remote_addr.ip(),
                     payload.signed_message.address,
                     req.uri(),
@@ -363,7 +306,7 @@ async fn router(
         AddressStatus::Blocked => {
             log::warn!(
                 "{}",
-                http_log_format!(
+                log_format!(
                     remote_addr.ip(),
                     payload.signed_message.address,
                     req_uri,
@@ -379,7 +322,7 @@ async fn router(
             if let Err(ProofOfFundingError::InvalidSignedMessage) = signed_message_status {
                 log::warn!(
                     "{}",
-                    http_log_format!(
+                    log_format!(
                         remote_addr.ip(),
                         payload.signed_message.address,
                         req_uri,
@@ -403,7 +346,7 @@ async fn router(
                 _ => {
                     log::warn!(
                         "{}",
-                        http_log_format!(
+                        log_format!(
                             remote_addr.ip(),
                             payload.signed_message.address,
                             req_uri,
@@ -418,7 +361,7 @@ async fn router(
                         Err(ProofOfFundingError::InsufficientBalance) => {
                             log::warn!(
                                 "{}",
-                                http_log_format!(
+                                log_format!(
                                     remote_addr.ip(),
                                     payload.signed_message.address,
                                     req_uri,
@@ -432,7 +375,7 @@ async fn router(
                         e => {
                             log::error!(
                                 "{}",
-                                http_log_format!(
+                                log_format!(
                                     remote_addr.ip(),
                                     payload.signed_message.address,
                                     req_uri,
@@ -450,7 +393,7 @@ async fn router(
             if db.rate_address(rate_limiter_key).await.is_err() {
                 log::error!(
                     "{}",
-                    http_log_format!(
+                    log_format!(
                         remote_addr.ip(),
                         payload.signed_message.address,
                         req_uri,
@@ -470,21 +413,6 @@ async fn router(
             .await
         }
     }
-}
-
-pub(crate) async fn serve(cfg: &'static AppConfig) -> GenericResult<()> {
-    let addr = format!("0.0.0.0:{}", cfg.port.unwrap_or(5000)).parse()?;
-
-    let router = make_service_fn(move |c_stream: &AddrStream| {
-        let remote_addr = c_stream.remote_addr();
-        async move { Ok::<_, GenericError>(service_fn(move |req| router(cfg, req, remote_addr))) }
-    });
-
-    let server = Server::bind(&addr).serve(router);
-
-    log::info!("AtomicDEX Auth API serving on http://{}", addr);
-
-    Ok(server.await?)
 }
 
 #[test]
@@ -556,25 +484,6 @@ fn test_respond_by_status() {
         let res = response_by_status(status_type).unwrap();
         assert_eq!(res.status(), status_type);
     }
-}
-
-#[test]
-fn test_get_real_address() {
-    let mut req = Request::new(Body::from(Vec::new()));
-
-    let addr = IpAddr::from_str("127.0.0.1").unwrap();
-    let socket_addr = SocketAddr::new(addr, 80);
-
-    let remote_addr = get_real_address(&req, &socket_addr).unwrap();
-    assert_eq!("127.0.0.1", remote_addr.ip().to_string());
-
-    req.headers_mut().insert(
-        HeaderName::from_static("x-forwarded-for"),
-        "0.0.0.0".parse().unwrap(),
-    );
-
-    let remote_addr = get_real_address(&req, &socket_addr).unwrap();
-    assert_eq!("0.0.0.0", remote_addr.ip().to_string());
 }
 
 #[tokio::test]
